adding corrections to styled components in last-saved.ts; converint nav.js to typescript
